richacl_equiv_mode(): Ignore owner-allowed permissions
The owner of a file is always allowed to change the file permissions and
the acl (ACE4_WRITE_ACL), take ownership of the file and set the owning
group to one of the groups the process is on (ACE4_WRITE_OWNER), and set
the mtime and atime (ACE4_WRITE_ATTRIBUTES).

When checking if an acl is equivalent to a file mode, allow the owner to
have these permissions explicitly (or not), and allow additional
permissions to be set in the EVERYONE@ ALLOW ace.

Signed-off-by: Andreas Gruenbacher <agruen@suse.de>