richacl_from_mode(): Add always-allowed and owner-allowed permissions
When synthesizing an acl from a file mode, explicitly include the
permissions which are always allowed, and which the owner is always
allowed.  This results in acls which more closely represent the actual
permissions granted.  Also, other systems may have different implied
permissions; such differences can now be made visible.

Signed-off-by: Andreas Gruenbacher <agruen@suse.de>