test: Added the -r option to resume a test at a given line
Added the -r option to resume a test at a given line.
It can be used with the -l option to run only a certain part of a test
or to skip part of a test.

Also slightly modified the behavior of the -l option to prevent running
one more command than intended if the line given wasn't a command.

Signed-off-by: Michael Groshans <groshans@umich.edu>
Signed-off-by: Aneesh Kumar K.V <aneesh.kumar@linux.vnet.ibm.com>