librichacl: Rename some of the richacl specific definitions
Some of the richacl.h definitions are richacl rather than nfs4 specific. Rename
those definitions to make the code more obvious.  Leave the ACL4_MASKED and
ACE4_SPECIAL_WHO definitions as they are though; otherwise bit flags with
different prefixes would have to be used for the same values which is very
confusing.