Also recognize nontrivial mode-equivalent acls
So far, richacl_equiv_mode() was relatively limited in the types of acl it
considered equivalent to a file mode: it only accepted masked acls with a
single everyone@:rwpxd::allow entry.

Change this to consider all acls equivalent to file modes if they only consist
of owner@, group@, and everyone@ entries and the owner@ permissions do not
depend on whether the owner is a member in the owning group.

With that, some of the test cases now no longer require an actual richacl; they
only use the file mode.