No longer add implicit owner permissions
The owner of a file is granted the write_attributes (A), write_acl (C), and
write_owner(o) permissions no matter if the file's acl grants those
permissions.  We were previously automatically adding those permissions when
owner@ entries were created on the basis that the acl should show the
permissions granted as well as possible.  However, those implicit permissions
are also granted when there are no owner@ entries, and so additional owner@
entries would sometimes have to be fabricated; if not, the acl wouldn't reflect
the actual permissions granted.  This seems worse than not adding implicitly
granted permissions in acls at all.