Update richacl_from_mode()
The kernel version of richacl_from_mode() now returns acls with inactive file
masks (i.e., the RICHACL_MASKED flag is not set); those acls are equivalent to
masked everyone@:rwpxdaAcCoS::allow acls with the file masks set to correspond
to the file mode.  Use the same code in librichacl.