librichacl: Switch to partial "write-through" semantics (1)
Change the permission check algorithm so that in masked acls, the owner is
always granted the owner mask permissions, and others are always granted the
other mask permissions: thus, upon chmod, the permissions of the owner and of
others is determined by the owner and other class permissions.  POSIX ACLs
behave the same way.

Setting the group class permissions and group mask do not set the permissions
of the owning group, they only define an upper bound.  This is also similar to
POSIX ACLs; the group mask corresponds to mask:: POSIX ACL entries.