Add support for unmapped identifiers
Allow identifiers (aka. principals) to be a special entity (OWNER@, GROUP@, or
EVERYONE@), a particular uid or gid, or be unmapped.  In the unmapped case,
the acl entry has the RICHACE_UNMAPPED_WHO flag set, and the identifier is a
null-terminated string.

Unmapped identifiers are useful for representing remote users and groups which
don't have a local uid / gid mapping.  They usually cannot be stored on local
file systems.

In the librichacl in-memory representation, unmapped identifiers are strings
allocated with malloc.  In the xattr representation, unmapped identifiers are
stored as a zero-separated and zero-terminated list of stings following the
acl.  In the text representation, the "u" flag indicates an unmapped
identifier.