Fix file creates with partial "write-through" semantics
When creating files, the create mode only defines an upper limit of the
permissions that may be granted, unlike with chmod, where the mode is "written
through" to the acl so that owner@ and everyone@ permissions can also be
increased.  Therefore, we need to treat these two cases differently
(distinguished by the new RICHACL_WRITE_THROUGH flag).

This also means that the owner mask can again contain more permissions than the
owner is actually allowed.  To avoid acl + mask combinations which cannot be
represented as acls , we need to reintroduce the group mask limit on
group-class entries.