Put user aces that match the owner into the owner class
Put user aces that match the current file owner into the owner class so that
creating with create mode 0600 will not mask permissions in those entries.
Otherwise, the owner would lose permissions from such entries on file create.

A chmod to mode 0600 will "write through" and create an owner@:rwp::allow ace,
so the owner would not be losing access in that case even without this
exception.