Add user: and group: prefixes in the text form
In the richacl text form, add user: and group: prefixes to the who
fields of regular users and groups and remove the g (identifier_group)
flag.  This makes acls easier to read and also matches how NFSv4 ACLs
are represented on other systems.

In addition, should we ever end up allowing entries which apply to users
and groups at the same time, to better represent CIFS ACLs or on systems
with User Private Groups, those could then be represented by omitting
that prefix.