richacl_inherit: Fix inheritance flags computation
Fix errors in computing the inheritance flags: for subdirectories, ACL
entries with the f and n flags and without the d flag must not be
inherited, and entries with the d flag must lose the i flag.  When the a
ACL flag is not set, the a flags of inherited entries must be cleared.

Add a test case enumerating all the inheritance flag combinations for
files and directories.